requirements.txt
alpaca-py==0.18.0
supabase==2.3.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
schedule==1.2.0
requests==2.31.0

